+++ packages/css-color-parser/src/functions/contrast-color.ts
@@ -8,6 +8,7 @@ import { Color } from '@csstools/color-helpers';
 import { XYZ_D50_to_sRGB_Gamut } from '../gamut-mapping/srgb';
 import { TokenType } from '@csstools/css-tokenizer';
 import { toLowerCaseAZ } from '../util/to-lower-case-a-z';
+import { contrast_ratio_wcag_2_1 } from '@csstools/color-helpers';
 
 export function contrastColor(colorMixNode: FunctionNode, colorParser: ColorParser): ColorData | false {
 	let backgroundColorData: ColorData | false = false;
@@ -56,8 +57,8 @@ export function contrastColor(colorMixNode: FunctionNode, colorParser: ColorPars
 		syntaxFlags: new Set([SyntaxFlag.ContrastColor, SyntaxFlag.Experimental]),
 	};
 
-	const contrastWhite = contrastRatio(backgroundColorData.channels, [1, 1, 1]);
-	const contrastBlack = contrastRatio(backgroundColorData.channels, [0, 0, 0]);
+	const contrastWhite = contrast_ratio_wcag_2_1(backgroundColorData.channels, [1, 1, 1]);
+	const contrastBlack = contrast_ratio_wcag_2_1(backgroundColorData.channels, [0, 0, 0]);
 
 	if (contrastWhite > contrastBlack) {
 		colorData.channels = [1, 1, 1];
@@ -66,26 +67,4 @@ export function contrastColor(colorMixNode: FunctionNode, colorParser: ColorPars
 	}
 
 	return colorData;
-}
-
-function luminance(color: Color): number {
-	// https://www.w3.org/TR/WCAG21/#dfn-relative-luminance
-	const [lumR, lumG, lumB] = color.map(component => {
-		return component <= 0.03928
-			? component / 12.92
-			: Math.pow((component + 0.055) / 1.055, 2.4);
-	});
-
-	return 0.2126 * lumR + 0.7152 * lumG + 0.0722 * lumB;
-}
-
-function contrastRatio(color1: Color, color2: Color): number {
-	// https://www.w3.org/TR/WCAG21/#dfn-contrast-ratio
-	const color1luminance = luminance(color1);
-	const color2luminance = luminance(color2);
-
-	const l1 = Math.max(color1luminance, color2luminance);
-	const l2 = Math.min(color1luminance, color2luminance);
-
-	return (l1 + 0.05) / (l2 + 0.05);
 }