+++ plugins/postcss-contrast-color-function/src/parse-contrast-color.ts
@@ -1,0 +1,42 @@
+import { ComponentValue, isCommentNode, isFunctionNode, isTokenNode, isWhitespaceNode } from '@csstools/css-parser-algorithms';
+import { TokenType } from '@csstools/css-tokenizer';
+
+const CONTRAST_COLOR_NAME_REGEX = /^contrast-color$/i;
+
+export function parseContrastColor(componentValue: ComponentValue): [ComponentValue, 'max'?] | false {
+	if (!isFunctionNode(componentValue) || !CONTRAST_COLOR_NAME_REGEX.test(componentValue.getName())) {
+		return false;
+	}
+
+	const meaningfulValues = componentValue.value.filter((value) => {
+		if (isWhitespaceNode(value) || isCommentNode(value)) {
+			return false;
+		}
+
+		return true;
+	});
+
+	if (meaningfulValues.length > 2) {
+		return false;
+	}
+
+	const color = meaningfulValues[0];
+	const modifier: ComponentValue | undefined = meaningfulValues[1];
+	if (!color) {
+		return false;
+	}
+
+	if (!modifier) {
+		return [color];
+	}
+
+	if (
+		!isTokenNode(modifier) ||
+		modifier.value[0] !== TokenType.Ident ||
+		modifier.value[4].value.toLowerCase() !== 'max'
+	) {
+		return false;
+	}
+
+	return [color, 'max'];
+}